Improve error handling in serial_read() (by Bernhard R. Link)
If read returns -1 and errno == EAGAIN that might mean there is no
data after we only got here when select claims there is, so warn about
something else eating our data. (But only do this if happening
multiple times in a row, to avoid bad luck with signals).

(This also resyncs ts41x's serial_read so that it no longer sets
buf[-1] to 0 on reading errors).

git-svn-id: http://qcontrol.googlecode.com/svn/trunk@37 cba486c8-e04f-0410-a7e3-79bda13a29fb