add systemd service/socket files
It is customary to ship the systemd files with the upstream package so
that the downstream distributions can simply use the files, they don’t
require changes downstream usually.

qcontrold.{service,socket} are for qcontrold and start qcontrold on
demand. qcontrol.service (without the d) overrides the init script with
the corresponding name and — implicitly — starts qcontrold by using the
statusled/powerled/buzzer commands on boot and shutdown.